sdk(js): rename experimental/sandbox -> sandbox (breaking)
BREAKING: The import path moves from `langsmith/experimental/sandbox`
to `langsmith/sandbox`. The old path no longer resolves.

- Move src/experimental/sandbox/ to src/sandbox/ (history preserved
  via git rename).
- Drop the `@experimental` JSDoc tags on `SandboxClient` and `Sandbox`
  and the experimental warning banner from src/sandbox/README.md.
- Update the package.json `exports` map, the `files` allowlist, the
  entrypoint script, the typedoc entry, and .gitignore to publish the
  new `./sandbox` entrypoint instead of `./experimental/sandbox`.
- Fix internal relative imports in src/sandbox/client.ts (`../../utils`
  -> `../utils`, etc.) and the imports in src/tests/sandbox*.test.ts.
- Refresh JSDoc `@example` blocks and README import statements to use
  the new path.

Verification: `pnpm build`, `pnpm lint` (no new warnings introduced),
`pnpm test:single src/tests/sandbox*.test.ts` (87/87 passing).

Migration: replace
  import { SandboxClient } from "langsmith/experimental/sandbox";
with
  import { SandboxClient } from "langsmith/sandbox";
