+++ js/scripts/create-entrypoints.js
@@ -30,7 +30,7 @@ const entrypoints = {
   "experimental/otel/processor": "experimental/otel/processor",
   "experimental/vercel": "experimental/vercel/index",
   "experimental/anthropic": "experimental/anthropic/index",
-  "experimental/sandbox": "experimental/sandbox/index",
+  sandbox: "sandbox/index",
 };
 
 const defaultEntrypoints = ["vitest/reporter"];
