+++ js/src/sandbox/command_handle.ts
