+++ js/src/sandbox/errors.ts
