+++ js/tsconfig.json
@@ -57,7 +57,7 @@
       "src/experimental/otel/processor.ts",
       "src/experimental/vercel/index.ts",
       "src/experimental/anthropic/index.ts",
-      "src/experimental/sandbox/index.ts"
+      "src/sandbox/index.ts"
     ]
   }
 }
